Filewalker and Timeline now pull from local repos and added Classroom states
