+++ browser/js/classroom/classroom.html
@@ -1,6 +1,4 @@
-  <!-- <div ng-controller='MainCtrl'>
-
-    <navbar></navbar>
+<div ng-controller='ClassroomCtrl'>
 
     <section class="main"> 
       <div class="directory">
@@ -33,5 +31,4 @@
       <p>made with<span class="glyphicon glyphicon-heart"></span>from team codestream</p>
     </div>
 
-</div> -->
-
+</div>
