+++ browser/js/common/directives/timeline/timeline.js
@@ -2,14 +2,14 @@
 app.directive('timeline', function() {
   return {
     restrict: 'E',
-    scope: {},
+    //scope: {},
     templateUrl: 'js/common/directives/timeline/timeline.html',
     controller: 'TimelineCtrl'
   };
 });
 
 app.controller('TimelineCtrl', function($scope, Timeline) {
-  Timeline.get(function(commits) {
+  Timeline.get($scope.room, function(commits) {
     $scope.commits = Timeline.sortByDate(commits);
   });
 })
