+++ browser/js/common/factories/TimeLineFactory.js
@@ -1,9 +1,9 @@
 app.factory('Timeline', function($http) {
   return {
-    get: function(cb) {
-      return $http.get('/api/timeline').then(function(res) {
-        cb(res.data);
-      });
+    get: function(dirname, cb) {
+       return $http.get('/api/timeline/' + dirname).then(function(res) {
+         cb(res.data);
+       });
     },
     sortByDate: function(commits) {
       return _.sortBy(commits, function(commit) {
