+++ package.json
@@ -8,7 +8,6 @@
     "postinstall": "./node_modules/bower/bin/bower install && gulp build"
   },
   "dependencies": {
-    "app-root-path": "^1.0.0",
     "bluebird": "^2.9.15",
     "body-parser": "^1.12.0",
     "bower": "^1.3.12",
