+++ server/app/routes/filetree/getfile.js
@@ -3,7 +3,7 @@ var fs = require('fs');
 module.exports = {
 
   getfile: function(path, cb) {
-  fs.readFile(path, 'utf8', cb); 
+  	fs.readFile(path, 'utf8', cb); 
   }
 
 };
