+++ server/app/routes/filetree/index.js
@@ -4,18 +4,20 @@ var getFile = require('./getfile');
 
 var router = require('express').Router();
 
-router.get('/', function (req, res) {
-    res.send(fileTree(__dirname + "/../../../"));
-});
-
 router.get('/file', function(req, res) {
-  var path = req.query.path;
-  getFile.getfile(path, function(err, file) {
+   var path = req.query.path;
+   getFile.getfile(path, function(err, file) {
     if(err) res.status(500).send(err);
     else {
       res.status(200).json({file: file }); 
     }
   }); 
 });
+
+router.get('/:dirname', function (req, res) {
+	var path = process.cwd() + '/repos/' + req.params.dirname;
+    res.send(fileTree(path));
+});
+
 
 module.exports = router;
