+++ server/app/routes/timeline/index.js
@@ -2,10 +2,10 @@
 var router = require('express').Router();
 var timeline = require('./timeline');
 
-router.get('/', function(req, res) {
-  timeline(function(commits) {
+router.get('/:dirname', function (req, res, next) {
+	timeline(req.params.dirname, function(commits) {
     res.status(200).send(commits);
   });
-});
+})
 
 module.exports = router;
