+++ server/app/views/index.html
@@ -25,6 +25,7 @@
     <script src="/main.js"></script>
 </head>
 <body ng-app="FullstackGeneratedApp">
+    <navbar></navbar>
     <div ui-view></div>
 </body>
 </html>
